Huge speed increase by converting findAudioDuplicate algorithm to C++
Add a FingerprintManager class in C++ that has a addSongAndCompare
method, accepting a songID, a fingerprint and a cancellation similarity
threshold. This method iterates over all songs already added to the
object's internal map/dictionary comparing them to the new added song.
Finally the song is added to the internal map and a list of tuples
is returned where each tuple is a matching song (songID, offset,
similarity).

This reduces the number of iterations/searches over the dictionary
of songs.

As a side effect, this fixes a problem with the original implementation
when two songs had exactly the same fingerprint, since the dictionary
was indexed by fingerprint, so only one of such songs was compared, thus
ignoring positive song matches with similarity = 1.0 .

Also, songs are added to the internal dictionary with a prepadding
of m_maxoffset zeros, so there's no need to manipulate data for each
comparison. In addition, the compareChromaprintFingerprintsAndOffset
method uses iterators to traverse the fingerprints (which is much
faster than index based accesses) and uses gcc's __builtin_popcount
since FingerprintManager already uses regular int values for
fingerprints.

If commit 8dffcdfa5511fe79db8b0a40bc7bfe6e3e8209f6 decreased
the time it took for a test run of 4759 songs from 215 seconds
to 126 seconds, this commit increases the speed quite a lot more so
the same test run takes 1.7 seconds.

Note: bard_ext.cpp uses C++17 and boost::python.

>>> p.print_stats()
Mon Aug  7 14:17:39 2017    profile_bard_cpp

         242517 function calls (235919 primitive calls) in 1.987 seconds

   Ordered by: cumulative time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    424/1    0.007    0.000    1.988    1.988 {built-in method builtins.exec}
        1    0.000    0.000    1.988    1.988 /usr/bin/bard:2(<module>)
        1    0.000    0.000    1.788    1.788 /usr/lib64/python3.6/site-packages/bard/bard.py:905(main)
        1    0.040    0.040    1.788    1.788 /usr/lib64/python3.6/site-packages/bard/bard.py:735(parseCommandLine)
        1    0.898    0.898    1.742    1.742 /usr/lib64/python3.6/site-packages/bard/bard.py:585(findAudioDuplicates2)
       55    0.643    0.012    0.643    0.012 {method 'execute' of 'sqlite3.Cursor' objects}
    451/2    0.001    0.000    0.200    0.100 <frozen importlib._bootstrap>:958(_find_and_load)
    451/2    0.001    0.000    0.200    0.100 <frozen importlib._bootstrap>:931(_find_and_load_unlocked)
    575/3    0.000    0.000    0.200    0.067 <frozen importlib._bootstrap>:197(_call_with_frames_removed)
    379/1    0.000    0.000    0.200    0.200 {built-in method builtins.__import__}
    438/2    0.001    0.000    0.199    0.099 <frozen importlib._bootstrap>:641(_load_unlocked)
    392/1    0.001    0.000    0.198    0.198 <frozen importlib._bootstrap_external>:672(exec_module)

Compare to the previous times:

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
    424/1    0.007    0.000  127.292  127.292 {built-in method builtins.exec}
        1    0.000    0.000  127.292  127.292 /usr/bin/bard:2(<module>)
        1    0.000    0.000  127.099  127.099 /home/antonio/git/antlarr/bard/bard/bard.py:847(main)
        1    0.128    0.128  127.099  127.099 /home/antonio/git/antlarr/bard/bard/bard.py:677(parseCommandLine)
        1    0.084    0.084  126.966  126.966 /home/antonio/git/antlarr/bard/bard/bard.py:509(findAudioDuplicates)
     3578    0.300    0.000  123.672    0.035 /home/antonio/git/antlarr/bard/bard/bard.py:105(compareChromaprintFingerprintsAndOffset2)
   354222   98.899    0.000  123.290    0.000 /home/antonio/git/antlarr/bard/bard/bard.py:38(compareChromaprintFingerprints)
299543702   24.271    0.000   24.271    0.000 {built-in method gmpy.popcount}
     3616    1.857    0.001    1.857    0.001 /home/antonio/git/antlarr/bard/bard/bard.py:526(<listcomp>)