Improvements to findAudioDuplicates
Now it's possible to specify one or more song ids and only recalculate
similarities for those songs. Also refactored a bit the
FingerprintManager class so the thresholds are now set for the instance
and a different threshold can be set for short songs (where the maximum
length of a "short" song can also be defined). This allows to be more
strict when comparing short songs.

In order to do this, we not store a vector of std::tuple instead of
std::pair in FingerprintManager