Initial version of the web interface
Lots of changes to implement a "web" command that
starts a web server that can be used to search the music
collection and stream audio. Currently the browser has
to support the audio format but in the future, automatic
media conversion will be done on the fly.

The web interface requires Werkzeug, Flask, Flask-Cors,
Flask-Login and Jinja2.

Also this commit adds a new "passwd" command that
changes the password for a user (by default the current user)
and if the user doesn't exist, also creates it.

This adds a dependency on the bcrypt python module.

This commit adds the jquery.js file, which is distributed
under the MIT license.