Use the ffmpeg libraries to decode audio instead of external libraries
This is a big change in the way bard decodes audio since it now uses
the ffmpeg libraries directly which means it's faster and has access to
a lot of information we can use and store in the database (like
container/stream bitrates, decode errors/warnings/messages, etc.).

This implementation of a decoder is done in a new bard_audiofile
extension that is easily used through the decodeAudio function
in utils.py .

The fixChecksums command has been modified to also use this new
implementation.

A new configuration option was added: 'enable_internal_checks' .
When enabled (set to true) then files are decoded twice: first
with bard_audiofile and then with pydub (which internally uses
the ffmpeg command) and the decoded raw data is compared to
check that the decoding worked fine.

This commit also introduces new tables in the database:
decode_properties, decode_messages and enumeration tables, which
are used to store integer numbers in columns instead of string values
(or lists of values). This makes it possible for example to store a
number identified the library versions used to decode a song, or using
numbers for sample formats. This is implemented in the DatabaseEnum
class in musicdatabase.py

Also, this commit adds a few columns to the database :
- insert_time and update_time columns to the songs table, to store the
  date a song was inserted in the database and the last time its
  information was updated
- insert_time and last_check_time columns to the checksums table.
- insert_time to the fingerprints table