Add new backup functionality to bard
Now bard can do smart backups of the music collection.

Why would you want to use bard instead of rsync or some similar tool?
For one important reason: the bard database makes bard smart.
First, let's see how to configure backups. In the config file
~/.config/bard a new section has to be created like:

    "backups" : {
        "hostname" : {
            "" : "username@hostname:/media/backups/music/"
            "/home/me/music2" : "username@hostname:/somewhere/else/music/"
            }

This defines a hostname where backups will be made when running
```bard backup hostname``` . When bard does such a backup, it'll copy
the music from all source paths to
username@hostname:/media/backups/music/ except the source path
/home/me/music2 which will be copied to
username@hostname:/somewhere/else/music/ .

Note that username@hostname has to be added to the ssh-agent so bard
can log in without asking for a password (this might be changed in the
future, but so far, that's a requirement).

So where's the smart in that? Well, when comparing the source and target
directories, bard will always ask before removing files from the target
directory, but it gives you more than y/n options. You can press ```c```
and check for differences. In that case, bard will download the files
that would be removed and compare against the files in the database,
reporting if the files are locally in a different location because
they've been moved even if the metadata has been changed since it'll
compare the audio streams specifically, not just the files themselves.

It's also possible to use ll or lr to ```list local/remote
directories``` or llr/lrr to ```list local/remote directories recursively```.