When comparing songs during backup, parse the local songs correctly.
The tags obtained from the local song is normalized, since it's obtained
from the database, while the tags obtained from the remote song is not
normalized, so when comparing the songs, this showed more changes than
actually were. To fix this, we now always read the file from disk, to
compare disk-metadata to disk-metadata.