First version of the MusicBrainz importer classes and database schema
This implements a first version of classes to download the MusicBrainz
schema/database dumps and import the pieces we want into the Bard
database (in a musicbrainz schema).

This is still work in progress and has to be optimized a lot in order
to use less memory and be faster before it can be used in Bard.