Let the user choose which file to move when two have the same checksum
Better support for the case of two files having the same checksum not
being in the previous location and now we find one of them in a new
location. We now let the user choose which one to move to the new
location.

For this, a new terminalkeyboard.py module gives the support of
working with the keyboard.