Add the new implementation of the musicbrainz importer script
Since the old implementation that tried to store some caches on memory
to access faster the records from musicbrainz tables on disk used
too much memory, this new implementation doesn't store anything on
memory and just focus on iterating FAST over the dump files and being
smart on filtering and returning generators that can be used to
iterate easily on the db dumps.

This has proved to work MUCH better and is going to be final
solution (for now, at least :)).

musicbrainz_database.py also includes some code used to get
information from the musicbrainz_database, that will be used by the
web interface.