Add the concept of Album
An Album is a MusicBrainz release instance. A single directory (optionally with
subdirectories) that contains files all belonging to an unique release.
There can be more than one album with the same release, but only one
release per album.

The table album_songs contains the relation between an album and its
songs, while the materialized views album_properties and album_releases
are automatically regenerated with the audio properties of each album
and the release associated to the album, respectively.

Note that an album may have more than one audio format within.