Fix some releases not being returned
There are releases that don't have the status set. Since
we were joining the query using that column with the
enum_release_status_values table, the releases having a null
value in the release_status were being ignored.
We better just return the status id and let someone interpret
the value somewhere else.