Retrieve more information about releases/albums and show them in the ui
Show the album audio properties next to each album as well as the
disambiguation text (honoring special tags like "usereleasecomment"
or "uselabel").

usereleasecomment = 1 means that the release disambiguation text should
be used.
usereleasecomment = 2 means that the comment tag should be used
usereleasecomment = 3 means that the release group tag should be used

uselabel = 1 means that the label name (which published the release)
should be used.
uselabel = 1 means that the label name and catalog number
should be used.

In case of more than one label/catalog numbers, the first release event
is used.