Add playlists support and much improved web ui
This commit includes support for playlists that are stored
in the database and are shown in a left panel in the web ui
with a splitter that can be dragged to expand/shrink the panel.

Also, there's a much nicer display of albums (and playlists)
and now songs can be dragged & dropped in playlists to add them.

Note there's not support to add playlists or remove songs from the web ui yet.
So those actions currently have to be done in the database directly.
Of course, I hope to change that soon if I get the time.

Also increased the jquery version from 3.3.1 to 3.4.1 .