Custom controls for media player and some refactoring of JS code
This commit adds a much better interface with custom controls
for the media player that match the rest of the page much better.

Also, it introduces the start of some badly needed refactoring of
the JS code. Now with some object oriented classes (and more to
come in the future).