Add support to read embedded cuesheets in flac audio files
Now flac files with embedded cuesheets (in a cuesheet block)
are read and the 'subtracks' are stored in a new cuesheets
table in the database. Since cuesheet blocks don't have a field
for track name, I'm using a custom tag SUBTRACKTITLES that
should contain a list of ordered titles for the tracks in the
cuesheet.