Add download_flags.py script to download country flags
This script was done originally by Luca Wehrstedt at
https://gist.githubusercontent.com/lw/3340802/raw/1a078a259fa93f6a20488086927fa4f48392bc6a/get_flags.py

It's not run or used directly by bard, but I modified and used it
to download all country flags (and the images were commited directly) so I'll keep
it here in case there's a need to run it again in the future.

My modifications will be submitted in the next commits.