Add multiselection support to the Chooser class
Chooser now allows the user to select multiple elements
by selecting/unselecting them using the space key.
When multiselection is enabled in the constructor the
chooser now returns the current element if the user
just pressed enter. If the user presses the space to
select one (or many) element(s), then pressing enter
doesn't return the current element, but the list of
selected elements