Add a GeneratedPlaylist class and support it everywhere.
A playlist of type `generated` is generated dynamically
from a generator string . Currently, only `shuffle` is supported
but the idea is to allow also things like `genre="Rock"` or
`rating>7 and genre="Pop"` for example (the exact format
still has to be defined).
Also, use a template for the add playlist dialog.