Add --rating, --my-rating and --others-rating parameters to ls/play commands
Add functionality to filter by ratings when listing or playing songs.
--my-rating filters just on ratings set by the current user.
--others-rating filters based on the average ratings of users other
than the current user.
--rating filters based on the processed rating of songs for the current
user. This means if the user has set a rating, it's used. Otherwise, if
one or more other users have set a rating, the average of those ratings
is used. In case no user has set a rating for a song, the song is
considered to have a default rating of 5.

Examples:
`bard ls --rating "> 8"`
`bard play --rating 10`