Fix finding the correct release group image for some cases
When a release group has an album/release which has different images
for each medium, then the image in the first medium directory
returned by the database query was used as the release group image.
We now try to first find an album image on any release from the
release group and only if none is found, we use an image from
a release medium.