Big refactor of database engine
This is a big refactor of how the database is handled in bard so now
sqlalchemy models are used to create all schemas and to
perform many of the queries. Also alembic is now used to handle
database schema versions so it will be possible to perform
automatic schema migrations when bard is run on an old database.

With this refactor, now it's possible to use sqlite and have
a fully working bard instance, including a working musicbrainz
integration.

The only drawback is that sqlalchemy has a bad support of views
(specially materialized views which I was using for album_release
and album_properties) so now instead of views I'm using regular
tables, removing all rows and inserting them back with a select
query on each "refresh". This is done this way because even if
it's possible to use materialized views using sqlalchemy_utils,
sqlite doesn't support materialized views, so I didn't want to
include special cases that would make schema migrations much
more difficult.