Add a preferred_locales config option instead of having it hardcoded
The preferred_locales config option is a list of locales (i.e. ['es', 'en'])
sorted by priority, so if an artist has different aliases in different locales
the first one is preferred to be used. This is useful for artists that use
scripts that the user doesn't usually work with, so "readable/writable"
artist names are shown by bard