Make init command write an example config file and refactor config system
Now running "bard init" writes an example config file that can be used
by the user to create an initial configuration easier.

Also, this changes how the configuration object is used across Bard.

The example configuration file is now located in the bard directory
and is installed using MANIFEST.in