Add compatibility with latest musicbrainz database schema
Now we allow to read UUID type values, support bigint columns
and also partition tables.

Also, link_partition_types are not ordered, so we have to be
careful when adding items so that parent types are imported
before child types. As a step in that direction, we now
import first types without a parent, which seems to be good
enough.