Add a new mb-check-redirected-uuids command
This command returns a list of albums that contain songs
with uuids that are obsoleted in the musicbrainz database
and are now redirected to new uuids.

The albums are displayed grouped in lists of 30 and in a
format that allows one to copy each line and pass it as
parameter to picard to easily reload and retag them.