Fix for ChannelLayout brokenness in libswresample
When setting up a SwrContext, the CodecCtx may be using
a ChannelLayout with order equal to AV_CHANNEL_ORDER_UNSPEC and
we can ask the SwrContext to use it as is, but the call to
swr_init then modifies it internally to set the default order (using
av_channel_layout_default as this change does).

The problem with that is that when decoding the data, frames
would still be using the AV_CHANNEL_ORDER_UNSPEC ch_layout but
the SwrContext expects that incoming frames use a channel layout
that matches what it stored internally, and there's no way that
matches, giving an error when resampling any frame.

Setting an ordered channel layout _BEFORE_ setting the channel
layout to use for the input frames fixes this.